requirements.txt
PySide6>=6.5.0
Pillow>=9.0.0
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.20
safetensors>=0.3.0
fpdf2>=2.7.0
numpy>=1.21.0
opencv-python>=4.7.0
